Move config to package `application/`
