+++ application/config.go
@@ -1,4 +1,4 @@
-package main
+package application
 
 import (
 	"log"
