+++ application/config_test.go
@@ -1,4 +1,4 @@
-package main
+package application
 
 import (
 	"os"
